+++ CHANGELOG.md
@@ -11,8 +11,18 @@ for further information about branching and tagging conventions.
 - Trim for input field 
 - Redeem default amount
 - Current company page
-- Skeleton loader for card
-- Global class for markdown
+- New component 'Card' for displaying all card
+- Skeleton loader for new component 'Card'
+- Global class for markdown `app__markdown`
+
+#### Removed
+- `card-viewer`
+- `asset-skeleton-loade`
+- `current-business-logo`
+- `current-business-indicator/index.vue`
+- `BusinessCardSkeleton`
+- `BusinessViewer`
+- `PromoCodeCardSkeleton`
 
 #### Changed
 - Moved logo and favicon to branding folder